volga/curel/requirements.txt
PyJWT
cryptography==3.3.2
gmpy2
